Separate dragDrop in a different module
